Consoidated python code to a single file
